requirements.txt
pandas==2.3.0
openpyxl==3.1.5
xlsxwriter==3.2.5
requests==2.32.4
streamlit==1.37.0
python-kucoin==2.2.0
